refactor(ai-adapters): extract chatStreamFromAdapter, the named BYOK arm
The daemon's BYOK arm built its `ChatStream` inline: bridge the worker's
`AbortSignal` onto an `AbortController`, then drive `chat({ adapter, ... })`.
ADR-0037 deferred extracting this until a second adapter -> `ChatStream` caller
existed. ADR-0038 trips that gate (the daemon's named BYOK builder is the second
caller), so the bridge moves into `chatStreamFromAdapter(adapter, systemPrompts)`
in the leaf, beside `createAdapterForModel`: the catalog gives the provider,
that builds the adapter, this turns it into the stream.

The leaf keeps two consumers and stays a leaf (ADR-0038); the `ChatStream`
signature is inlined here, the same decoupling `createEpicenterProviderChatStream`
makes, so the leaf still does not depend on the workspace core. The hosted route
is not a caller (it returns an HTTP Response, not a `ChatStream`), so it stays
inline.
